pyproject.toml
[tool.poetry]
name = "cat-classifier"
version = "0.1.0"
description = "Learning project for image classification using neural networks"
authors = ["marcosandoval <kada9001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.1.0"
matplotlib = "^3.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
